refactor: 응답 헤더 타입 변경
