+++ tomcat/src/main/java/org/apache/coyote/http11/request/RequestURI.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import java.io.File;
 import java.io.IOException;
