+++ tomcat/src/main/java/org/apache/coyote/http11/response/ResponseHeaders.java
@@ -1,6 +1,6 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.response;
 
-import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -8,12 +8,8 @@ import java.util.stream.Collectors;
 public class ResponseHeaders {
     private final Map<String, String> headers;
 
-    public ResponseHeaders(final Map<String, String> headers) {
-        this.headers = headers;
-    }
-
     public ResponseHeaders() {
-        this.headers = new HashMap<>();
+        this.headers = new LinkedHashMap<>();
     }
 
     public void add(final String key, final String value) {
